# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: API per il servizio di messaggistica "WASAText" del corso WASA.
  version: "1.0.0"
servers:
  - url: /
...
security: # Regole di sicurezza globali
  - bearerAuth: [] # Definito in components/securitySchemes/bearerAuth
paths:
...
  /settings/username:
    put:
      tags: ["settings"] 
      summary: Set my username
      description: Aggiorna il nome utente dell'utente autenticato.
      operationId: setMyUsername
      requestBody:
        description: Il nuovo username desiderato.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetMyUsernameRequest'
      responses:
        '200':
          description: Username aggiornato con successo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '409':
          $ref: '#/components/responses/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /groups: 
    post:
      tags: ["groups"]
      summary: Create a new group
      description: Crea una nuova conversazione di gruppo includendo l'utente autenticato e gli utenti specificati.
      operationId: createGroup 
      requestBody:
        description: Nome del gruppo e lista degli ID dei membri da includere (oltre all'utente autenticato).
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: Gruppo creato con successo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /conversations/{convId}/name:
    put:
      tags: ["groups"]
      summary: Set group name
      description: Aggiorna il nome di una conversazione di gruppo specificata. L'utente autenticato deve essere membro del gruppo.
      operationId: setGroupName
      parameters:
        - $ref: '#/components/parameters/ConvID'
      requestBody:
        description: Il nuovo nome desiderato per il gruppo.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetGroupNameRequest'
      responses:
        '200':
          description: Nome del gruppo aggiornato con successo. Restituisce la conversazione aggiornata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /conversations/{convId}/photo:
    put:
      tags: ["groups"]
      summary: Set group photo
      description: Aggiorna l'URL della foto di una conversazione di gruppo specificata. L'utente autenticato deve essere membro del gruppo.
      operationId: setGroupPhoto
      parameters:
        - $ref: '#/components/parameters/ConvID'
      requestBody:
        description: Il nuovo URL della foto del gruppo.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetPhotoRequest'
      responses:
        '200':
          description: Foto del gruppo aggiornata con successo. Restituisce la conversazione aggiornata.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Conversation'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /conversations/{convId}/members:
    post:
      tags: ["groups"]
      summary: Add a user to a group
      description: Aggiunge un utente specificato (tramite ID) a una conversazione di gruppo. L'utente autenticato deve essere membro del gruppo.
      operationId: addToGroup
      parameters:
        - $ref: '#/components/parameters/ConvID'
      requestBody:
        description: L'ID dell'utente da aggiungere al gruppo.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserIdRequest'
      responses:
        '204': 
          description: Membro aggiunto con successo. Nessun contenuto restituito.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /conversations/{convId}/members/me:
    delete:
      tags: ["groups"] 
      summary: Leave a group
      description: Permette all'utente autenticato di abbandonare una conversazione di gruppo specificata.
      operationId: leaveGroup
      parameters:
        - $ref: '#/components/parameters/ConvID'
      responses:
        '204':
          description: Gruppo abbandonato con successo. Nessun contenuto restituito.
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '401':
          $ref: '#/components/responses/Unauthorized'
components:
  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "string"
      description: "L'identificativo utente (ottenuto dal login) da usare come Bearer token nell'header `Authorization: Bearer <identifier>`."
  parameters:
    ConvID:
      name: convId
      in: path
      required: true
      description: ID univoco della conversazione.
      schema:
        $ref: '#/components/schemas/InternalID' 
...
  responses:
    BadRequest: # 400
      description: Formato dei dati non valido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized: # 401
      description: Autenticazione richiesta o token non valido.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden: # 403
      description: Non hai i permessi per eseguire questa azione.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound: # 404
      description: La risorsa richiesta non è stata trovata.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict: # 409
      description: La risorsa esiste già o è in conflitto.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError: # 500
      description: Errore interno del server.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UserID:
      description: ID univoco in formato UUID (usato per utenti e bearer token).
      type: string
      pattern: '^[a-fA-F0-9]{8}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{4}-[a-fA-F0-9]{12}$'
      minLength: 36
      maxLength: 36
      example: "abcdef01-2345-6789-abcd-ef0123456789"
    InternalID:
      description: ID univoco generato dal sistema (per conversazioni, messaggi, reazioni).
      type: string
      pattern: '^.+$'
      minLength: 1
      maxLength: 50
      example: "msg-789"
    Username:
      description: Nome univoco scelto dall'utente.
      type: string
      pattern: '^.+$'
      minLength: 3
      maxLength: 16
      example: "Maria"
    PhotoURL:
      description: URL di un'immagine (foto profilo o messaggio).
      type: string
      format: uri
      pattern: '^.+$'
      minLength: 1
      maxLength: 2048
      example: "https://example.com/photos/image.jpg"
...
    GroupName:
      description: Nome di una conversazione o di un gruppo.
      type: string
      pattern: '^.+$'
      minLength: 1
      maxLength: 50
      example: "Gruppo Lavoro"
...
    User:
      description: Rappresenta un utente registrato nel sistema.
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UserID'
        username:
          $ref: '#/components/schemas/Username'
        photoUrl:
          $ref: '#/components/schemas/PhotoURL'
      required: ["id", "username"]
...
    ErrorResponse:
      description: Schema generico per le risposte di errore.
      type: object
      properties:
        message:
          type: string
          description: Messaggio descrittivo dell'errore.
          minLength: 1
          maxLength: 256
          pattern: '^.+$'
          example: "Si è verificato un errore."
      required: ["message"]
...
    SetMyUsernameRequest:
      description: Dati inviati per aggiornare il nome utente.
      type: object
      properties:
        username:
          $ref: '#/components/schemas/Username'
      required: ["username"]
    SetPhotoRequest:
      description: Dati inviati per aggiornare l'URL della foto profilo dell'utente o del gruppo.
      type: object
      properties:
        photoUrl:
          $ref: '#/components/schemas/PhotoURL'
      required: ["photoUrl"]
    UserIdRequest: 
      description: Un oggetto che contiene un singolo ID utente (UUID).
      type: object
      properties:
        userId:
          $ref: '#/components/schemas/UserID'
      required: ["userId"]
    CreateGroupRequest:
      description: Dati inviati per creare una nuova conversazione di gruppo.
      type: object
      properties:
        groupName:
          $ref: '#/components/schemas/GroupName'
        memberIds:
          type: array
          description: Lista degli ID (univoci) degli utenti da aggiungere al gruppo (oltre all'utente che crea).
          minItems: 1 
          maxItems: 99 
          items:
            $ref: '#/components/schemas/UserID'
      required: ["groupName", "memberIds"]
    SetGroupNameRequest:
      description: Dati inviati per rinominare un gruppo.
      type: object
      properties:
        name:
          $ref: '#/components/schemas/GroupName'
      required: ["name"]
...